// src/app/services/swagger/model/sharedBattleReport.ts
/**
 * BoF REST API
 * Battle for honor interface
 *
 * OpenAPI spec version: 0.1.4
 * Contact: webmaster@batleforhonor.de
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { AbstractId } from './abstractId';
import { Alliance } from './alliance';

/**
 * .
 */
export interface SharedBattleReport {
    /**
     * .
     */
    idBattleReport: number;
    /**
     * .
     */
    participatingUsers: Array<AbstractId>;
    /**
     * .
     */
    sharedWithAlliance: Array<Alliance>;
    /**
     * .
     */
    sharedWithUsers: Array<AbstractId>;
    /**
     * .
     */
    shareWithEveryone: boolean;
}